Add the "include_functor" extension.
This extension adds the structure item form "include functor F", where F must be
a functor whose parameter can be "filled in" with the previous contents of the
current module.  This eliminates a common source of boilerplate where modules
create submodules that duplicate some of their contents and exist only for the
purpose of applying a functor.  For example:

    module Foo = struct
      module T = struct
        type t = int [@@deriving compare]
      end

      include T
      include Comparable.Make(T)
    end

Can now be written:

    module Foo = struct
      type t = int [@@deriving compare]

      include functor Comparable.Make
    end

Where the only difference is that the latter Foo doesn't have the submodule T.