Allow use of `local_` to adjust regions for loops.
This patch changes the local stack region behavior for the body of for loops,
and the body and condition of while loops.  Now, the user may request that no
new stack region is created for these positions, and instead local allocations
happen in the region containing the loop.

This is of limited use in isolation, but is useful with the coming "let mutable"
patch, where we want to allow locals in a loop body to "escape" to mutable
variables declared in the scope containing the loop.