Add the "let mutable" extension.
This extension allows let-bound variables to be marked as mutable, as in:

let add_55 y =
  let mutable x = y in
  for i = 1 to 10 do
    x <- x + i
  done;
  x

Mutable variables are like "ref"s, but don't give you a way to refer to the ref
cell itself - all uses refer to the stored value.

The typechecker ensures that mutable variables do not escape the scope in which
they are defined.  This static check ensures, roughly, that the variable can
be implemented efficiently as a register, without allocation.  Additionally, it
allows for local values to be stored in mutable variables.