Use proper destructors for extension-enabled ASTs
The goal here is to more structurally avoid the possibility of
forgetting about extensions. This doesn't really actually get us
there, though, because we can't change Parsetree to avoid using
e.g. pexp_desc. And even if we could make that change, we might
not want to, because it's just too convenient to just use pexp_desc.
Until we have active patterns of some sort, we may be unable to
improve upon the status quo.

However, as I was working this, I identified a few spots where I
think there are live bugs for lack of looking for extensions. And
I think having a destructor, as this commit adds, is a positive little
improvement.